+++ packages/organizations/origin-organization-irec-api/src/connection/dto/connection.dto.ts
@@ -27,6 +27,15 @@ export class ConnectionDTO extends ShortConnectionDTO {
     @ApiProperty({ type: String })
     userName: string;
 
+    @ApiProperty({ type: String })
+    clientId: string;
+
+    @ApiProperty({ type: String })
+    clientSecret: string;
+
+    @ApiProperty({ type: Boolean })
+    active: boolean;
+
     public static wrap(connection: Connection): ConnectionDTO {
         return plainToClass(ConnectionDTO, connection);
     }