+++ packages/ui/origin-ui-core/src/components/certificates/DetailView/CertificateDetailView.tsx
@@ -16,6 +16,7 @@ import { useOriginConfiguration } from '../../../utils/configuration';
 import { DeviceDTO } from '@energyweb/origin-device-registry-irec-form-api-client';
 import { fromGeneralSelectors } from '../../../features';
 import { ClaimDataDTO } from '@energyweb/issuer-api-client';
+import moment from 'moment';
 
 interface IProps {
     id: number;
@@ -100,9 +101,12 @@ export function CertificateDetailView(props: IProps) {
 
                     break;
                 case 'TransferSingle':
-                    if (event._from === '0x0000000000000000000000000000000000000000') {
+                    if (event.from === '0x0000000000000000000000000000000000000000') {
                         label = t('certificate.event.name.initialOwner');
                         description = transformAddress(event.to);
+                    } else if (event.to === '0x0000000000000000000000000000000000000000') {
+                        label = '';
+                        description = '';
                     } else {
                         label = t('certificate.event.name.changedOwnership');
                         description = t('certificate.event.description.transferred', {
@@ -146,11 +150,14 @@ export function CertificateDetailView(props: IProps) {
                     requestor: transformAddress(request.owner),
                     amount: EnergyFormatter.format(request.energy, true)
                 }),
-                timestamp: request.created
+                timestamp: moment((request as any).createdAt).unix()
             });
         }
 
-        setEvents(deduplicate(resolvedEvents).sort((a, b) => a.timestamp - b.timestamp));
+        const filteredEvents =
+            resolvedEvents?.filter((event) => !!event.label || !!event.description) || [];
+
+        setEvents(deduplicate(filteredEvents).sort((a, b) => a.timestamp - b.timestamp));
     }
 
     const [certificateData, setCertificateData] = useState<Array<TCertificateData[]>>([]);